Remove `FormMixin` plugin code and expand regular type
