+++ mypy_django_plugin/transformers/forms.py
@@ -1,8 +1,7 @@
-from mypy.plugin import ClassDefContext, MethodContext
-from mypy.types import CallableType, Instance, NoneTyp, TypeType, get_proper_type
-from mypy.types import Type as MypyType
+from mypy.nodes import TypeInfo
+from mypy.plugin import ClassDefContext
 
-from mypy_django_plugin.lib import helpers
+from mypy_django_plugin.lib import fullnames, helpers
 
 
 def make_meta_nested_class_inherit_from_any(ctx: ClassDefContext) -> None:
@@ -14,40 +13,10 @@ def make_meta_nested_class_inherit_from_any(ctx: ClassDefContext) -> None:
         meta_node.fallback_to_any = True
 
 
-def get_specified_form_class(object_type: Instance) -> TypeType | None:
-    form_class_sym = object_type.type.get("form_class")
-    if form_class_sym:
-        form_class_type = get_proper_type(form_class_sym.type)
-        if isinstance(form_class_type, CallableType):
-            return TypeType(form_class_type.ret_type)
-    return None
+def transform_form_class(ctx: ClassDefContext) -> None:
+    sym = ctx.api.lookup_fully_qualified_or_none(fullnames.BASEFORM_CLASS_FULLNAME)
+    if sym is not None and isinstance(sym.node, TypeInfo):
+        bases = helpers.get_django_metadata_bases(sym.node, "baseform_bases")
+        bases[ctx.cls.fullname] = 1
 
-
-def extract_proper_type_for_get_form(ctx: MethodContext) -> MypyType:
-    object_type = ctx.type
-    assert isinstance(object_type, Instance)
-
-    form_class_type = get_proper_type(helpers.get_call_argument_type_by_name(ctx, "form_class"))
-    if form_class_type is None or isinstance(form_class_type, NoneTyp):
-        form_class_type = get_specified_form_class(object_type)
-
-    if isinstance(form_class_type, TypeType) and isinstance(form_class_type.item, Instance):
-        return form_class_type.item
-
-    if isinstance(form_class_type, CallableType):
-        form_class_ret_type = get_proper_type(form_class_type.ret_type)
-        if isinstance(form_class_ret_type, Instance):
-            return form_class_ret_type
-
-    return ctx.default_return_type
-
-
-def extract_proper_type_for_get_form_class(ctx: MethodContext) -> MypyType:
-    object_type = ctx.type
-    assert isinstance(object_type, Instance)
-
-    form_class_type = get_specified_form_class(object_type)
-    if form_class_type is None:
-        return ctx.default_return_type
-
-    return form_class_type
+    make_meta_nested_class_inherit_from_any(ctx)
