+++ tests/typecheck/views/generic/test_edit.yml
@@ -82,8 +82,8 @@
         class MyCreateView(CreateView[Article, ArticleModelForm]):
             def some(self) -> None:
                 reveal_type(self.get_form())  # N: Revealed type is "main.ArticleModelForm"
-                reveal_type(self.get_form(SubArticleModelForm))  # N: Revealed type is "main.SubArticleModelForm"
-                reveal_type(self.get_form(AnotherArticleModelForm))  # N: Revealed type is "main.AnotherArticleModelForm"  # E: Argument 1 to "get_form" of "FormMixin" has incompatible type "type[AnotherArticleModelForm]"; expected "type[ArticleModelForm] | None"  [arg-type]
+                reveal_type(self.get_form(SubArticleModelForm))  # N: Revealed type is "main.ArticleModelForm"
+                reveal_type(self.get_form(AnotherArticleModelForm))  # N: Revealed type is "main.ArticleModelForm"  # E: Argument 1 to "get_form" of "FormMixin" has incompatible type "type[AnotherArticleModelForm]"; expected "type[ArticleModelForm] | None"  [arg-type]
     installed_apps:
         - myapp
     files:
